Minor refactor - Consolidate credentials and config into one file - Update @types/node - Better imports of configs
