--- config/credentials.example.json
@@ -1,4 +1,0 @@
-{
-    "DiscordToken": "BotToken",
-    "MongoDB": "mongodb://localhost:27017"
-}
