+++ src/lib/checkDomain.ts
@@ -1,11 +1,11 @@
 import { Config } from "../types/config.js";
 import untypedConfig from "../../config/config.json" assert { type: "json" };
 
-const settings = untypedConfig as Config;
+const { Domains } = untypedConfig as Config;
 
 export async function isAllowedDomain(domain: string): Promise<boolean> {
     let out = false;
-    if (settings.Domains.some((v) => domain.includes(v))) {
+    if (Domains.some((v) => domain.includes(v))) {
         await fetch(domain)
             .then((r) => {
                 out = r.ok;
